minor
git-svn-id: svn+ssh://astrometry.net/svn/trunk/src/astrometry@9037 03a3cea6-2c03-0410-ac9b-9271b1c66c29

+++ util/tic.c
@@ -36,7 +36,7 @@ double timenow() {
         ERROR("Failed to get time of day");
         return -1.0;
     }
-    return (tv.tv_sec - 3600*24*365*30) + tv.tv_usec * 1e-6;
+    return (double)(tv.tv_sec - 3600*24*365*30) + tv.tv_usec * 1e-6;
 }
 
 double millis_between(struct timeval* tv1, struct timeval* tv2) {
